+++ server/run.py
@@ -3,7 +3,7 @@
 appインスタンスを呼び出し、Flaskサーバーを起動
 """
 
-from app.views import app
+from app import app
 
 if __name__ == "__main__":
     app.run(debug=True)
